+++ checker/tests/nullness-records/BasicRecord.java
@@ -1,0 +1,14 @@
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+// @below-java16-jdk-skip-test
+public record BasicRecord(String str) {
+
+  public static BasicRecord makeNonNull(String s) {
+    return new BasicRecord(s);
+  }
+
+  public static BasicRecord makeNull(@Nullable String s) {
+    // :: error: argument
+    return new BasicRecord(s);
+  }
+}